requirements.txt
Flask==2.1.1
Flask-SocketIO==5.1.1
python-dotenv==0.20.0
redis==4.1.1
Flask-Session==0.4.0